Добавлена обработка файлов, пока без вывода, но в базу сохраняюся

+++ src/main/java/com/project/database_2b2t_org_ru/dao/ThreadRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThreadRepository extends JpaRepository<Thread, Integer> {
+public interface ThreadRepository extends JpaRepository<Thread, Long> {
 
     Page<Thread> findAll(Pageable pageable);
 }
